@@ -1,6 +1,6 @@
 {
-  "name": "lottie-react-web",
-  "version": "2.1.4",
+  "name": "@jadbox/lottie-react-web",
+  "version": "2.1.15",
   "description": "Lottie wrapper for React on the web.",
   "types": "./index.d.ts",
   "repository": {
@@ -21,37 +21,42 @@
     "postpublish": "npm run publish-storybook"
   },
   "devDependencies": {
-    "@kadira/storybook": "^2.35.3",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-react-app": "^0.2.1",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.4.0",
+    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-transform-runtime": "^7.4.0",
+    "@babel/polyfill": "^7.4.0",
+    "@babel/preset-env": "^7.4.2",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/register": "^7.4.0",
+    "@babel/runtime": "^7.4.2",
+    "@storybook/react": "^5.0.5",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.5",
     "chai": "^3.5.0",
     "enzyme": "^2.2.0",
-    "eslint": "^3.6.0",
-    "eslint-config-airbnb": "^12.0.0",
-    "eslint-plugin-import": "^1.16.0",
-    "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-react": "^6.3.0",
+    "eslint": "^5.15.3",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-react": "^7.12.4",
     "git-url-parse": "^6.0.1",
     "jsdom": "^9.5.0",
     "mocha": "^5.2.0",
-    "prop-types": "^15.6.1",
-    "react": "^15.6.2",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.5",
     "react-addons-test-utils": "^15.6.2",
-    "react-dom": "^15.6.2",
+    "react-dom": "^16.8.5",
     "sinon": "^1.17.6",
-    "standard-version": "^4.4.0"
+    "standard-version": "^4.4.0",
+    "webpack": "^4.29.6",
+    "lottie-web": "^5.4.3"
   },
   "peerDependencies": {
-    "react": "^0.14.7 || ^15.0.0 || ^16.0.0"
+    "react": "^0.14.7 || ^15.0.0 || ^16.0.0",
+    "lottie-web": "^5.4.3"
   },
   "dependencies": {
-    "babel-runtime": "^6.26.0",
-    "lottie-web": "^5.2.1",
     "lottie-api": "bodymovin/lottie-api#master"
   },
   "main": "dist/index.js",
