@@ -3,7 +3,7 @@
 // This is an auto generated file with React CDK.
 // Do not modify this file.
 
-import { configure } from '@kadira/storybook';
+import { configure } from '@storybook/react';
 
 function loadStories() {
   require('../src/stories');
